+++ utils/forest_shim/Cargo.toml
@@ -18,9 +18,9 @@ fvm_ipld_encoding3 = { package = "fvm_ipld_encoding", version = "0.3" }
 fvm_shared.workspace = true
 fvm_shared3 = { workspace = true, features = ["proofs", "arb"] }
 num-bigint.workspace = true
-num.workspace = true
-num-traits.workspace = true
 num-derive.workspace = true
+num-traits.workspace = true
+num.workspace = true
 quickcheck.workspace = true
 serde = { workspace = true, features = ["derive"] }
