+++ vm/state_migration/src/nv18/eth_account.rs
@@ -10,6 +10,7 @@ use forest_shim::{
 use forest_utils::db::BlockstoreExt;
 use fvm_ipld_blockstore::Blockstore;
 
+/// Creates the Ethereum Account actor in the state tree.
 pub fn create_eth_account_actor<BS: Blockstore + Clone + Send + Sync>(
     store: &BS,
     actors_out: &mut StateTree<BS>,
